Move firebase functions out of index
